// frontend/src/backend.d.ts
export interface SocialMediaLinks {
    instagram?: string;
    whatsapp?: string;
    discord?: string;
    youtube?: string;
}